we mostly have input working now ....

+++ src/java/maprunner/util/Hex.java
@@ -13,8 +13,8 @@ public class Hex {
         
         for( byte b : input ) {
 
-            int i = b + 128;
-
+            int i = ((int)b) + 128;
+            
             buff.append( String.format( "%02x ", i ) );
 
         }
